+++ src/voeventparse/tests/test_voeventparse.py
@@ -211,7 +211,7 @@ class TestWhereWhen(TestCase):
                           observatory_location=vp.definitions.observatory_location.geosurface)
         self.assertTrue(vp.valid_as_v2_0(self.v))
         self.assertEqual(self.coords1, vp.pull_astro_coords(self.v))
-        self.assertIsNotNone(vp.pull_isotime(self.v))
+        self.assertIsNotNone(vp.pull_event_time_as_utc(self.v))
         astrocoords = self.v.xpath(
             'WhereWhen/ObsDataLocation/ObservationLocation/AstroCoords'
         )[0]