+++ src/voeventparse/voevent.py
@@ -10,6 +10,10 @@ import voeventparse.definitions
 voevent_v2_0_schema = etree.XMLSchema(
     etree.fromstring(voeventparse.definitions.v2_0_schema_str))
 
+from ._version import get_versions
+
+__version__ = get_versions()['version']
+
 
 def Voevent(stream, stream_id, role):
     """Create a new VOEvent element tree, with specified IVORN and role.
@@ -49,8 +53,11 @@ def Voevent(stream, stream_id, role):
     etree.SubElement(v, 'Who')
     etree.SubElement(v, 'What')
     etree.SubElement(v, 'WhereWhen')
-    v.Who.Description = ('VOEvent created with voevent-parse: '
-                         + 'https://github.com/timstaley/voevent-parse')
+    v.Who.Description = (
+        'VOEvent created with voevent-parse, version {}. '
+        'See https://github.com/timstaley/voevent-parse for details.').format(
+        __version__
+    )
     return v